Separate link and tracking handler
for better test handling
